add users app and sorting filter
