+++ core/urls.py
@@ -22,6 +22,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('shop.urls', namespace='shop')),
     path('products/', include('products.urls', namespace='products')),
-    path('accounts/', include('django.contrib.auth.urls')),
+    path('user/', include('users.urls', namespace='users')),
 
 ]
