+++ shop/views.py
@@ -1,5 +1,3 @@
-from django.contrib.auth import logout
-from django.contrib.auth.decorators import login_required
 from django.shortcuts import render
 from products.models import Category
 
@@ -13,16 +11,4 @@ def about(request):
     context={"title":"about" }
     return render(request, "shop/about.html",context)
 
-def login_user(request):
-    return render(request, "registration/login.html")
 
-
-@login_required
-def profile_view(request):
-    return render(request, "shop/profile.html")
-
-
-@login_required
-def user_logout(request):
-    logout(request)
-    return render(request, "registration/logout.html", {})
