+++ users/templates/users/login.html
@@ -1,0 +1,15 @@
+{% extends "shop/base.html" %}
+{% load static %}
+
+{% block css %}
+	
+{% endblock  %}
+
+{% block content %}	
+					<div class="auto-container">
+						<h2> FORM LOGIN </h2>
+
+								
+
+					</div>
+{% endblock%}
